# tye-talk-08-replication/api.weather/WeatherForecast.cs
using System;

namespace api.weather
{
    public class WeatherForecast
    {
        public string PostalCode { get; set; }

        public DateTime Date { get; set; }

        public int TemperatureC { get; set; }

        public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);

        public string Summary { get; set; }
    }
}